Move CommercialPaper into asset subpackage
